split layout info and associated classes into a separate file, fix headers
